% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.PSAboot.R
\name{as.data.frame.PSAbootSummary}
\alias{as.data.frame.PSAbootSummary}
\title{Convert the results of PSAboot summary to a data frame.}
\usage{
\method{as.data.frame}{PSAbootSummary}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{results of \code{\link{summary.PSAboot}}}

\item{row.names}{row names.}

\item{optional}{unused.}

\item{...}{unused.}
}
\value{
a data.frame.
}
\description{
Convert the results of PSAboot summary to a data frame.
}
