% Generated by roxygen2 (4.0.1): do not edit by hand
\name{balance.matchit}
\alias{balance.matchit}
\title{Returns balance for each covariate using \code{\link{matchit}}}
\usage{
balance.matchit(mi, covs)
}
\arguments{
\item{mi}{the results of \code{\link{matchit}}}

\item{covs}{data frame or matrix of covariates. Factors should already be recoded.
See \code{\link{cv.trans.psa}}}
}
\value{
a named vector with one element per covariate.
}
\description{
Returns balance for each covariate using \code{\link{matchit}}
}

