% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSAboot.R
\name{getPSAbootMethods}
\alias{getPSAbootMethods}
\title{Returns a vector with the default methods used by \code{\link{PSAboot}}.}
\usage{
getPSAbootMethods()
}
\value{
a vector of methods for use by \code{\link{PSAboot}}
}
\description{
The current default methods are:
\describe{
\item{Stratification}{\code{\link{boot.strata}}}
\item{ctree}{\code{\link{boot.ctree}}}
\item{rpart}{\code{\link{boot.rpart}}}
\item{Matching}{\code{\link{boot.matching}}}
\item{MatchIt}{\code{\link{boot.matchit}}}
}
}
\details{
The default methods can be changed by setting the \code{PSAboot.methods} option
using \code{options('PSAboot.methods'=c(...))} where \code{...} is a named
list of functions.
}
