% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PSAboot.matrix.R
\name{matrixplot}
\alias{matrixplot}
\title{Matrix Plot of Boostrapped Propensity Score Analysis}
\usage{
matrixplot(bm)
}
\arguments{
\item{bm}{result from \code{\link{PSAboot}}.}
}
\description{
Matrix Plot of Boostrapped Propensity Score Analysis
}

