% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PSAboot.R
\name{plot.PSAboot}
\alias{plot.PSAboot}
\title{Plot the results of PSAboot}
\usage{
\method{plot}{PSAboot}(x, sort = "all", ci.sig.color = "red",
  plot.overall = FALSE, plot.bootstrap = TRUE, ...)
}
\arguments{
\item{x}{result of \code{\link{PSAboot}}.}

\item{sort}{how the sort the rows by mean difference. Options are to sort
using the mean difference from matching, stratificaiton, both 
individually, or no sorting.}

\item{ci.sig.color}{the color used for confidence intervals that do not span zero.}

\item{plot.overall}{whether to plot vertical lines for the overall (non-bootsrapped)
estimate and confidence interval.}

\item{plot.bootstrap}{whether to plot vertical lines for the bootstrap pooled
estimate and confidence interval.}

\item{...}{currently unused.}
}
\description{
Plot the results of PSAboot
}

