\name{print.PSAbootSummary}
\alias{print.PSAbootSummary}
\title{Print method for PSAboot Summary.}
\usage{
  \method{print}{PSAbootSummary} (x, digits = 3, ...)
}
\arguments{
  \item{x}{result of \code{\link{summary.PSAboot}}}

  \item{digits}{desired number of digits after the decimal
  point.}

  \item{...}{unused.}
}
\description{
  Print method for PSAboot Summary.
}

