\name{psa.strata}
\alias{psa.strata}
\title{Propensity Score Analysis using Stratification}
\usage{
  psa.strata(Y, Tr, strata, trim = 0)
}
\arguments{
  \item{Y}{response variable.}

  \item{Tr}{treatment variable.}

  \item{strata}{strata identifier.}

  \item{trim}{allows for a trimmed mean as outcome measure,
  where trim is from 0 to .5 (.5 implying median).}
}
\description{
  Propensity Score Analysis using Stratification
}

