\docType{package}
\name{PSAboot-package}
\alias{PSAboot-package}
\title{An R Package for Propensity Score Analysis}
\description{
  Bootstrapping procedures for Propensity Score Analysis.
}
\author{
  Jason Bryer \email{jason@bryer.org}
}
\keyword{analysis}
\keyword{bootstrapping}
\keyword{graphics}
\keyword{propensity}
\keyword{psa}
\keyword{score}

