\docType{package}
\name{PSAgraphics2-package}
\alias{PSAgraphics2}
\alias{PSAgraphics2-package}
\title{An R Package for Propensity Score Analysis}
\description{
  This packages provides functions to estimate and
  visualize multilvel propensity score analysis.
}
\details{
  This package extends the principles put forth by the
  \code{PSAgraphics} (Helmreich, Pruzek, & Xiong, 2010).
}
\author{
  Jason Bryer \email{jason@bryer.org}
}
\references{
  \url{http://cran.r-project.org/web/packages/PSAgraphics/PSAgraphics.pdf}
  \url{http://www.jstatsoft.org/v29/i06/}
}
\seealso{
  \code{PSAgraphics}
}
\keyword{analysis}
\keyword{graphics}
\keyword{propensity}
\keyword{psa}
\keyword{score}

