% Generated by roxygen2 (4.0.1): do not edit by hand
\name{q25}
\alias{q25}
\title{Return the 25th percentile.}
\usage{
q25(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{numeric vector.}

\item{na.rm}{logical; if true, any NA and NaN's are removed from x before the
quantiles are computed}

\item{...}{other parameters passed to \code{\link{quantile}}.}
}
\value{
the 25th percetile.
}
\description{
Return the 25th percentile.
}

