\name{summary.bootmatch}
\alias{summary.bootmatch}
\title{Summary of pooled results from bootmatch}
\usage{
  \method{summary}{bootmatch} (object, digits = 3, ...)
}
\arguments{
  \item{object}{result of \code{\link{bootmatch}}.}

  \item{...}{currently unused.}

  \item{digits}{desired number of digits after the decimal
  point.}
}
\value{
  a list with pooled summary statistics.
}
\description{
  Summary of pooled results from bootmatch
}

