\name{summary.PSAboot}
\alias{summary.PSAboot}
\title{Summary of pooled results from bootmatch}
\usage{
  \method{summary}{PSAboot} (object, ...)
}
\arguments{
  \item{object}{result of \code{\link{bootmatch}}.}

  \item{...}{currently unused.}
}
\value{
  a list with pooled summary statistics.
}
\description{
  Summary of pooled results from bootmatch
}

