% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.matching.R
\name{balance.matching}
\alias{balance.matching}
\title{Returns balance for each covariate from propensity score matching.}
\usage{
balance.matching(index.treated, index.control, covs)
}
\arguments{
\item{index.treated}{a vector with the index of treated rows in \code{covs}.}

\item{index.control}{a vector with the index of control rows in \code{covs}.}

\item{covs}{data frame or matrix of covariates. Factors should already be recoded.
See \code{\link{cv.trans.psa}}}
}
\value{
a named vector with one element per covariate.
}
\description{
This function is function is primarily used by [PSAboot::balance()] and
probably does not need to be called directly.
}
