% Generated by roxygen2 (4.0.1): do not edit by hand
\name{boot.matching}
\alias{boot.matching}
\title{Matching package implementation for bootstrapping.}
\usage{
boot.matching(Tr, Y, X, X.trans, formu, estimand = "ATE", ...)
}
\arguments{
\item{estimand}{character string for estimand, either ATE, ATT, or ATC. See
\code{\link{Match}} for more details.}

\item{...}{other parameters passed to \code{\link{Match}}.}

\item{Tr}{vector indicating treatment assignment.}

\item{Y}{vector of outcome.}

\item{X}{matrix or data frame of covariates.}

\item{X.trans}{a data frame of \code{X} with factors recoded. See \code{\link{cv.trans.psa}}}

\item{formu}{the formula to use to estimate propensity scores. Note that the
dependent varaible (i.e. treatment variable) name will be updated using
the \code{Tr} vector.}
}
\description{
Matching package implementation for bootstrapping.
}

