\name{boot.matchit}
\alias{boot.matchit}
\title{MatchIt package implementation for bootstrapping.}
\usage{
  boot.matchit(Tr, Y, X, ...)
}
\arguments{
  \item{Tr}{vector indicating treatment assignment.}

  \item{Y}{vector of outcome.}

  \item{X}{matrix or data frame of covariates.}

  \item{...}{other parameters passed from
  \code{\link{PSAboot}}}
}
\description{
  MatchIt package implementation for bootstrapping.
}

