% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hist.PSAboot}
\alias{hist.PSAboot}
\title{Histogram of PSA bootstrap results}
\usage{
\method{hist}{PSAboot}(x, ...)
}
\arguments{
\item{x}{result of \code{\link{PSAboot}}.}

\item{...}{other parameters passed to \code{\link{geom_histogram}}}
}
\description{
Histogram of PSA bootstrap results
}

