% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.PSAboot.R
\name{hist.PSAboot}
\alias{hist.PSAboot}
\title{Histogram of PSA bootstrap results}
\usage{
\method{hist}{PSAboot}(x, ...)
}
\arguments{
\item{x}{result of \code{\link{PSAboot}}.}

\item{...}{other parameters passed to \code{\link{geom_histogram}}}
}
\description{
Histogram of PSA bootstrap results
}

