% Generated by roxygen2 (4.0.1): do not edit by hand
\name{matrixplot}
\alias{matrixplot}
\title{Matrix Plot of Boostrapped Propensity Score Analysis}
\usage{
matrixplot(bm)
}
\arguments{
\item{bm}{result from \code{\link{PSAboot}}.}
}
\description{
Matrix Plot of Boostrapped Propensity Score Analysis
}

