% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PSAboot.matrix.R
\name{matrixplot}
\alias{matrixplot}
\title{Matrix Plot of Bootstrapped Propensity Score Analysis}
\usage{
matrixplot(bm)
}
\arguments{
\item{bm}{result from \code{\link{PSAboot}}.}
}
\value{
Nothing returned. Creates a plot using the [graphics::pairs()] function.
}
\description{
Matrix Plot of Bootstrapped Propensity Score Analysis
}
