% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.PSAboot}
\alias{print.PSAboot}
\title{Print results of PSAboot}
\usage{
\method{print}{PSAboot}(x, ...)
}
\arguments{
\item{x}{result of \code{\link{PSAboot}}.}

\item{...}{currently unused.}
}
\description{
Print results of PSAboot
}

